# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface KnowledgeItem {
  id: number;
  title: string;
  answer: string;
  category: string;
  tags: string[];
  isPremium: boolean;
  updatedAt: string;
}

export type AccountRole = typeof AccountRole[keyof typeof AccountRole];


export const AccountRole = {
  assistant: 'assistant',
  admin: 'admin',
} as const;

export type AccountStatus = typeof AccountStatus[keyof typeof AccountStatus];


export const AccountStatus = {
  active: 'active',
  blocked: 'blocked',
} as const;

export interface Account {
  accountNumber: string;
  displayName: string;
  username: string;
  role: AccountRole;
  isPremium: boolean;
  status: AccountStatus;
  createdAt: string;
  lastSeenAt: string;
}

export type AccountRegistrationRole = typeof AccountRegistrationRole[keyof typeof AccountRegistrationRole];


export const AccountRegistrationRole = {
  assistant: 'assistant',
  admin: 'admin',
} as const;

export interface AccountRegistration {
  /** @minLength 2 */
  displayName: string;
  /** @minLength 2 */
  username: string;
  role?: AccountRegistrationRole;
  adminCode?: string;
}

export interface AccountUpdate {
  /** @minLength 2 */
  displayName?: string;
  /** @minLength 2 */
  username?: string;
}

export type AccountAccessUpdateRole = typeof AccountAccessUpdateRole[keyof typeof AccountAccessUpdateRole];


export const AccountAccessUpdateRole = {
  assistant: 'assistant',
  admin: 'admin',
} as const;

export type AccountAccessUpdateStatus = typeof AccountAccessUpdateStatus[keyof typeof AccountAccessUpdateStatus];


export const AccountAccessUpdateStatus = {
  active: 'active',
  blocked: 'blocked',
} as const;

export interface AccountAccessUpdate {
  isPremium?: boolean;
  role?: AccountAccessUpdateRole;
  status?: AccountAccessUpdateStatus;
}

export interface DashboardStats {
  totalKnowledge: number;
  commandCount: number;
  priceCount: number;
  premiumCount: number;
  totalAccounts: number;
  activeAccounts: number;
}

export type ListKnowledgeParams = {
search?: string;
category?: string;
};