# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Live Russia Assistant Hub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: knowledge
  - name: accounts
  - name: admin
  - name: dashboard
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: Search the Live Russia knowledge base
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Matching knowledge entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    get:
      operationId: getKnowledge
      tags: [knowledge]
      summary: Get a knowledge entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Knowledge entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
        "404":
          description: Knowledge entry not found
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /accounts/register:
    post:
      operationId: registerAccount
      tags: [accounts]
      summary: Register an assistant account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRegistration"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid registration
  /accounts/{accountNumber}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get an account by its static number
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update the current account
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
  /admin/accounts:
    get:
      operationId: listAccounts
      tags: [admin]
      summary: List all assistant accounts
      responses:
        "200":
          description: Assistant accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
  /admin/accounts/{accountNumber}/access:
    patch:
      operationId: updateAccountAccess
      tags: [admin]
      summary: Change Premium or account role
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountAccessUpdate"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Account not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    KnowledgeItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        answer:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        isPremium:
          type: boolean
        updatedAt:
          type: string
      required: [id, title, answer, category, tags, isPremium, updatedAt]
    Account:
      type: object
      properties:
        accountNumber:
          type: string
        displayName:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [assistant, admin]
        isPremium:
          type: boolean
        status:
          type: string
          enum: [active, blocked]
        createdAt:
          type: string
        lastSeenAt:
          type: string
      required:
        - accountNumber
        - displayName
        - username
        - role
        - isPremium
        - status
        - createdAt
        - lastSeenAt
    AccountRegistration:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
        username:
          type: string
          minLength: 2
        role:
          type: string
          enum: [assistant, admin]
        adminCode:
          type: string
      required: [displayName, username]
    AccountUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 2
        username:
          type: string
          minLength: 2
    AccountAccessUpdate:
      type: object
      properties:
        isPremium:
          type: boolean
        role:
          type: string
          enum: [assistant, admin]
        status:
          type: string
          enum: [active, blocked]
    DashboardStats:
      type: object
      properties:
        totalKnowledge:
          type: integer
        commandCount:
          type: integer
        priceCount:
          type: integer
        premiumCount:
          type: integer
        totalAccounts:
          type: integer
        activeAccounts:
          type: integer
      required:
        - totalKnowledge
        - commandCount
        - priceCount
        - premiumCount
        - totalAccounts
        - activeAccounts