# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountRole } from './accountRole';
import type { AccountStatus } from './accountStatus';

export interface Account {
  accountNumber: string;
  displayName: string;
  username: string;
  role: AccountRole;
  isPremium: boolean;
  status: AccountStatus;
  createdAt: string;
  lastSeenAt: string;
}