# lib/api-zod/src/generated/types/accountAccessUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountAccessUpdateRole } from './accountAccessUpdateRole';
import type { AccountAccessUpdateStatus } from './accountAccessUpdateStatus';

export interface AccountAccessUpdate {
  isPremium?: boolean;
  role?: AccountAccessUpdateRole;
  status?: AccountAccessUpdateStatus;
}