# lib/api-zod/src/generated/types/accountRegistration.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountRegistrationRole } from './accountRegistrationRole';

export interface AccountRegistration {
  /** @minLength 2 */
  displayName: string;
  /** @minLength 2 */
  username: string;
  role?: AccountRegistrationRole;
  adminCode?: string;
}