# lib/api-zod/src/generated/types/accountUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AccountUpdate {
  /** @minLength 2 */
  displayName?: string;
  /** @minLength 2 */
  username?: string;
}