# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalKnowledge: number;
  commandCount: number;
  priceCount: number;
  premiumCount: number;
  totalAccounts: number;
  activeAccounts: number;
}