# lib/api-zod/src/generated/types/knowledgeItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Live Russia Assistant Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgeItem {
  id: number;
  title: string;
  answer: string;
  category: string;
  tags: string[];
  isPremium: boolean;
  updatedAt: string;
}